requirements.txt
Flask==2.0.3
psycopg2-binary==2.8.6
Nginx==1.21.6
Flask-SQLAlchemy==2.5.1